pyproject.toml
[project]
name = "experiment"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "nidaqmx>=1.3.0",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages=["experiment"]

[tool.hatch.build.targets.sdist]
include=[
    "*.py",
    "*.html",
    "*.css",
    "*.js"
]

[dependency-groups]
dev = [
    "ipython>=9.6.0",
]
